LPS-55762 Move classes to correct package
