+++ modules/apps/nested-portlets/nested-portlets-web/src/com/liferay/nested/portlets/web/settings/internal/NestedPortletsPortletInstanceSettingsIdMapping.java
@@ -12,10 +12,11 @@
  * details.
  */
 
-package com.liferay.nested.portlets.web.configuration.internal;
+package com.liferay.nested.portlets.web.settings.internal;
 
 import com.liferay.nested.portlets.web.configuration.NestedPortletsPortletInstanceConfiguration;
-import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
+import com.liferay.nested.portlets.web.constants.NestedPortletsPortletKeys;
+import com.liferay.portal.kernel.settings.definition.SettingsIdMapping;
 
 import org.osgi.service.component.annotations.Component;
 
@@ -23,12 +24,17 @@ import org.osgi.service.component.annotations.Component;
  * @author Juergen Kappler
  */
 @Component
-public class NestedPortletsPortletInstanceConfigurationBeanDeclaration
-	implements ConfigurationBeanDeclaration {
+public class NestedPortletsPortletInstanceSettingsIdMapping
+	implements SettingsIdMapping {
 
 	@Override
-	public Class getConfigurationBeanClass() {
+	public Class<?> getConfigurationBeanClass() {
 		return NestedPortletsPortletInstanceConfiguration.class;
+	}
+
+	@Override
+	public String getSettingsId() {
+		return NestedPortletsPortletKeys.NESTED_PORTLETS;
 	}
 
 }
